Add signature signing outline
