+++ Cargo.toml
@@ -45,6 +45,7 @@ url = "2.5.0"
 
 sha2 = "0.10.8"
 hmac = "0.12.1"
+rsa = "0.9.7"
 rand_chacha = "0.3.1"
 rand = "0.8.5"
 webpki = "0.22.4"
